+++ android/src/main/java/io/github/mertguner/sound_generator/SoundGenerator.java
@@ -14,7 +14,6 @@ import io.github.mertguner.sound_generator.generators.triangleGenerator;
 import io.github.mertguner.sound_generator.handlers.isPlayingStreamHandler;
 import io.github.mertguner.sound_generator.models.WaveTypes;
 
-@TargetApi(Build.VERSION_CODES.CUPCAKE)
 public class SoundGenerator {
 
     private Thread bufferThread;
@@ -23,7 +22,12 @@ public class SoundGenerator {
     private boolean isPlaying = false;
     private int minSamplesSize;
     private WaveTypes waveType = WaveTypes.SINUSOIDAL;
-    private float rightVolume = 1, leftVolume = 1;
+    private float rightVolume = 1, leftVolume = 1, volume = 1, dB = -20;
+    private boolean cleanStart = false;
+
+    public void setCleanStart(boolean cleanStart) {
+        this.cleanStart = cleanStart;
+    }
 
     public void setAutoUpdateOneCycleSample(boolean autoUpdateOneCycleSample) {
         if (generator != null)
@@ -68,12 +72,38 @@ public class SoundGenerator {
     }
 
 
-    public void setVolume(float volume) {
+    public void setVolume(float volume, boolean recalculateDecibel) {
         volume = Math.max(0, Math.min(1, volume));
+        this.volume = volume;
+
+        if(recalculateDecibel) {
+            if (volume >= 0.000001f) {
+                this.dB = 20f * (float) Math.log10(volume);
+            } else {
+                this.dB = -120f;
+            }
+        }
 
         if (audioTrack != null) {
             audioTrack.setStereoVolume(leftVolume * volume, rightVolume * volume);
         }
+    }
+
+    public float getVolume() {
+        return volume;
+    }
+
+    public void setDecibel(float dB) {
+        this.dB = dB;
+        float lineerVolume = (float) Math.pow(10f, (dB / 20f) );
+        if (lineerVolume < 0.000001f) {
+            lineerVolume = 0f;
+        }
+        setVolume(lineerVolume, false);
+    }
+
+    public float getDecibel() {
+        return dB;
     }
 
     public void setWaveform(WaveTypes waveType) {
@@ -124,6 +154,11 @@ public class SoundGenerator {
         if (bufferThread != null || audioTrack == null) return;
 
         isPlaying = true;
+
+        if (cleanStart) {
+            generator.resetFrequency();
+            generator.updateOnce();
+        }
 
         bufferThread = new Thread(new Runnable() {
             @Override